backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
supabase==2.9.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Machine Learning
scikit-learn==1.5.2
numpy==2.1.2
pandas==2.2.3
joblib==1.4.2

# Computer Vision & OCR
opencv-python-headless==4.10.0.84
pillow==10.4.0
ultralytics==8.3.197

# Data Processing
pydantic==2.9.2
pydantic-settings==2.6.0
python-dateutil==2.9.0.post0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Monitoring & Logging
loguru==0.7.2

# Task Scheduling
APScheduler==3.10.4

# Development
pytest==8.3.3
pytest-asyncio==0.24.0